Move evolution functions into main namespace
