+++ fwdpy11/src/evolve_population/init.cc
@@ -5,12 +5,11 @@ namespace py = pybind11;
 
 void init_no_stopping(py::module &);
 void init_evolve_with_tree_sequences(py::module &);
-void init_evolve_without_tree_sequences(py::module & m);
+void init_evolve_without_tree_sequences(py::module &m);
 
-PYBIND11_MODULE(_evolve_population, m)
+void
+init_evolution_functions(py::module &m)
 {
-    m.doc() = "Evolution under a Wright-Fisher model using tree sequences.";
-
     init_no_stopping(m);
     init_evolve_with_tree_sequences(m);
     init_evolve_without_tree_sequences(m);
