+++ tests/test_pickling.py
@@ -111,7 +111,7 @@ class testPickleDiploidPopulation(unittest.TestCase):
 class testPickleDiploidPopulationTreeSequences(unittest.TestCase):
     @classmethod
     def setUp(self):
-        import fwdpy11.wright_fisher_ts
+        import fwdpy11
         import numpy as np
         self.N = 1000
         self.demography = np.array([self.N]*self.N, dtype=np.uint32)
@@ -134,7 +134,7 @@ class testPickleDiploidPopulationTreeSequences(unittest.TestCase):
         self.params = fwdpy11.ModelParams(**self.p)
         self.rng = fwdpy11.GSLrng(101*45*110*210)
         self.pop = fwdpy11.DiploidPopulation(self.N, 1.0)
-        fwdpy11.wright_fisher_ts.evolve(self.rng, self.pop, self.params, 100)
+        fwdpy11.evolvets(self.rng, self.pop, self.params, 100)
 
     def testPickleNodeTable(self):
         import pickle
